apps/ai-assistant/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
openai==1.30.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
httpx==0.27.0
tenacity==8.3.0
pillow==10.3.0
pytesseract==0.3.10
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
pypdf==4.2.0
